Product KPI GUI Sample Python Project
This Program was originally used for institutional research. Please ignore raw code variable names. 
This program utilizes excel sheet "ConsumerDummyData". This program is an example of a user interface in which the user can check KPI's of a particular product. This program generates reportings as well visuals needed for that product KPI's. For this example the program generates reports on how well a product is reaching consumer compared to the product line and brand totals.